Rename files in common/CreateJoinForm
